Back end chat



See merge request !74

+++ frontend/app/js/core/services/httpInterceptors.service.js
@@ -35,17 +35,19 @@
         }
 
         function responseError(response) {
-            var toast = $injector.get('toastService');
-            var reg = /^20+.$/;
-            var isStatusOk = reg.test(response.status);
-            var message = response.data && response.data.message ? ': ' + response.data.message : '';
+            if (response.data && response.data.message) {
+                var toast = $injector.get('toastService');
+                var reg = /^20+.$/;
+                var isStatusOk = reg.test(response.status);
+                var message = response.data.message;
 
-            if (response.status > 0 && !isStatusOk) {
-                if (response.status === 401 || response.status === 403) {
-                    localService.removeAuth();
+                if (response.status > 0 && !isStatusOk) {
+                    if (response.status === 401 || response.status === 403) {
+                        localService.removeAuth();
+                    }
+
+                    toast.error(message);
                 }
-
-                toast.error('Server error' + message);
             }
 
             return $q.reject(response);
